+++ modelmapper/src/main/java/com/rokt/modelmapper/data/DataBinding.kt
@@ -11,6 +11,7 @@ import com.rokt.modelmapper.uimodel.ResponseOptionModel
 import com.rokt.modelmapper.utils.CURRENT_POSITION_PLACEHOLDER
 import com.rokt.modelmapper.utils.TOTAL_OFFERS_PLACEHOLDER
 import com.rokt.modelmapper.utils.transformToAnchorTag
+import java.util.TreeMap
 
 interface DataBinding {
     fun bindValue(value: String, contextKey: String? = null, offerModel: OfferModel?, itemIndex: Int = 0): BindData
@@ -115,6 +116,17 @@ private class PlaceholderReplacer(
                             }
                         }
 
+                        CREATIVE_IMAGE_NAMESPACE -> {
+                            val sanitisedKey = getSanitisedDataKey(key)
+                            val imageVal = offer?.creative?.images?.get(sanitisedKey.substringBeforeLast('.'))
+                            if (imageVal != null) {
+                                result =
+                                    imageVal.properties.get<String>(
+                                        TypedKey<String>(sanitisedKey.substringAfterLast('.')),
+                                    )
+                            }
+                        }
+
                         CATALOG_ITEM_NAMESPACE -> {
                             val catalogItemCopy = getCatalogItemCopy(key, itemIndex)
                             if (catalogItemCopy != null) {
@@ -174,6 +186,17 @@ private class PlaceholderReplacer(
         offer?.creative?.links?.get(getSanitisedDataKey(key))?.transformToAnchorTag()
 }
 
+internal fun getOfferImages(inputKey: String = "", offerModel: OfferModel?): Map<Int, OfferImageModel> {
+    val result = TreeMap<Int, OfferImageModel>()
+    for ((key, value) in offerModel?.creative?.images ?: emptyMap()) {
+        if (key.startsWith(inputKey)) {
+            val suffix = key.substring(inputKey.length + 1) // Extract the number after the prefix
+            result[suffix.toInt()] = value
+        }
+    }
+    return result
+}
+
 private enum class TemplateDataPrefix(val value: String) {
     DATA("DATA"),
     STATE("STATE"),
@@ -190,5 +213,6 @@ private const val CREATIVE_RESPONSE_NAMESPACE = "creativeResponse"
 private const val CREATIVE_COPY_NAMESPACE = "creativeCopy"
 private const val CREATIVE_LINKS_NAMESPACE = "creativeLink"
 private const val CATALOG_ITEM_NAMESPACE = "catalogItem"
+private const val CREATIVE_IMAGE_NAMESPACE = "creativeImage"
 private val INDICATOR_POSITION = listOf("IndicatorPosition", "indicatorPosition")
 private val TOTAL_OFFERS = listOf("TotalOffers", "totalOffers")