+++ modelmapper/src/main/java/com/rokt/modelmapper/uimodel/LayoutSchemaUiModel.kt
@@ -199,6 +199,19 @@ sealed class LayoutSchemaUiModel(
         val transition: WhenUiTransition,
         val hide: WhenUiHidden?,
     ) : LayoutSchemaUiModel()
+
+    class DataImageCarouselUiModel(
+        ownModifiers: ImmutableList<StateBlock<ModifierProperties>>?,
+        containerProperties: ImmutableList<StateBlock<ContainerProperties>>?,
+        conditionalTransitionModifiers: ConditionalTransitionModifier?,
+        val images: Map<Int, ImageUiModel>,
+        val duration: Long,
+        val indicator: ProgressIndicatorItemUiModel?,
+        val activeIndicator: ProgressIndicatorItemUiModel?,
+        val seenIndicator: ProgressIndicatorItemUiModel?,
+        val progressIndicatorContainer: ProgressIndicatorItemUiModel?,
+        val customStateKey: String,
+    ) : LayoutSchemaUiModel(ownModifiers, containerProperties, conditionalTransitionModifiers)
 }
 
 sealed class ButtonUiModel(
@@ -506,6 +519,7 @@ enum class BooleanWhenUiCondition {
 enum class OpenLinks {
     Internally,
     Externally,
+    Passthrough,
 }
 
 data class WhenUiTransition(